#11 Add Windows Support (#24)
* Update build rule to allow building on Windows

* Move main and main_wnd to win folder

* Move Linux/main.cc to serverless/

* Update Readme

* Sync compatibility with Windows and Linux in readme

* Refined readme

* Fix crash


* Fix bug


* Remove useless modification


* Polish code

* Remove git-lfs

* Update ONNXInfer

* Remove LFS from Azure Pipelines

* readd so files

* Remove media files

* Remove gitattributes and to include test media files

* Polish Readme and code

* delete onnx-model.onnx

* add onnx-model without lfs
