--- examples/peerconnection/serverless/corpus/.gitattributes
@@ -1,3 +1,0 @@
-*.onnx filter=lfs diff=lfs merge=lfs -text
-*.wav filter=lfs diff=lfs merge=lfs -text
-*.yuv filter=lfs diff=lfs merge=lfs -text
