--- examples/peerconnection/serverless/corpus/.gitignore
@@ -1,2 +1,0 @@
-*.yuv
-*.wav
